// src/typing/types/ICheckoutValues.ts
export type IValues = {
  Nome: string,
  Cpf: string
  Whatsapp: string,
  Estado: string,
  Cidade: string,
  Endereco: string,
  Complemento: string,
  Numero: string,
  Cep: string,
  Bairro: string
  email: string
  CardNumber: string
  CardExpire: string
  CardName: string
  CardCVC: string
  parcelas: number
}